requirements.txt
ansible==1.9.4
appdirs==1.4.3
asgi-ipc==1.3.1
asgiref==1.1.1
asn1crypto==0.22.0
attrs==16.3.0
autobahn==0.18.1
Automat==0.5.0
bleach==2.0.0
cffi==1.10.0
channels==1.1.3
constantly==15.1.0
cryptography==1.8.1
daphne==1.2.0
Django==1.10.7
django-cors-headers==2.0.2
django-treebeard==4.1.0
djangorestframework==3.6.2
djangorestframework-recursive==0.1.1
enum34==1.1.6
html5lib==0.999999999
idna==2.5
incremental==16.10.1
ipaddress==1.0.18
Jinja2==2.9.6
MarkupSafe==1.0
msgpack-python==0.4.8
MySQL-python==1.2.5
packaging==16.8
paramiko==2.1.2
posix-ipc==1.0.0
pyasn1==0.2.3
pycparser==2.17
pycrypto==2.6.1
pyOpenSSL==16.2.0
pyparsing==2.2.0
pytz==2017.2
PyYAML==3.12
six==1.10.0
Twisted==17.1.0
txaio==2.6.2
webencodings==0.5.1
zope.interface==4.3.3
